refactor: done due to ECMAScript modules
